requirements.txt
numpy==1.19.5
lmdb
pyarrow
pyyaml
easydict
ConfigArgParse
tensorboard
einops
matplotlib
tqdm
torchsnooper
ema-pytorch
accelerate
soundfile
pandas==1.3.4
transforms3d
scipy
scikit-learn
librosa
omegaconf
sox
rich
ffmpeg-normalize
blobfile
pydub 
praat-parselmouth 
essentia 
TextGrid 
h5py
ipdb
timm